requirements.txt
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic>=1.11
fastapi>=0.111
uvicorn>=0.30
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3.0
pydantic>=2.0
supertokens-python>=0.29.0